+++ CHANGELOG.md
@@ -4,9 +4,11 @@ English | [中文](./CHANGELOG.zh.md)
 
 All notable changes to this project will be documented in this file.
 
-## 0.2.0 - 2026-02-01
+## 0.3.0 - 2026-02-01
 
 ### Features
+- Implement Gantt-like Event Timeline for Stage details using ECharts
+- Support uncompressed Spark Standalone event logs
 - Enhance stage summary metrics to match Spark UI parity
 
 ## 0.1.0 - 2026-02-01